pyproject.toml
[project]
name = "rpyframe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "imageio>=2.37.0",
    "pillow>=11.2.1",
    "pygame>=2.6.1",
    "screeninfo>=0.8.1",
]
